aes: Split FFI helpers into their own submodule.
Make it clearer which types are used in the FFI and start
separating the (unsafe) code for dealing with the C/assembly
implementations from the (mostly safe) rest of the code.

`git difftool HEAD^1:src/aead/aes.rs src/aead/aes/ffi.rs`.
